Replaced `ToString()` with `TryFormat` and stack allocated spans
The `ToString()` method used for string conversion has been replaced with a `TryFormat` method utilizing stack allocated spans. Two new constants, `sizeOfSignificand` and `sizeOfExponent`, have been added to define the size of the new `Span<char>` named `inputSpan`, which is allocated on the stack using `stackalloc`. The `TryFormat` method is now used to format the input into the `inputSpan`, with the number of characters written stored in `charsWritten`. The `integerComponent` variable now references the `inputSpan` instead of the string representation of the input, aligning with the previous TODO comment.