Updated assertion methods and added new utility methods
Updated the `CreateFromFloatingPoint<TFloat>` and `CreateFromInteger<TInteger>` methods in the `SignificantNumber` struct to use a new assertion method, `AssertDoesImplementGenericInterface`. This method checks if the provided type implements the specified generic interface. Also added a new method `DoesImplementGenericInterface` that returns a boolean value indicating whether a given type implements a specified generic interface. Two additional utility methods, `EnsureTypeIsInterface` and `EnsureTypeIsGeneric`, have been added to throw an `ArgumentException` if the provided type is not an interface or a generic type, respectively.