Updated error handling and refactored `SignificantNumber` struct
This commit includes several updates to the `SignificantNumber` struct. The `CompareTo(object? obj)` method now provides a more detailed error message when an argument exception is thrown, specifying the type of the object that is not comparable with `SignificantNumber`.

The `DoesImplementGenericInterface(Type type, Type genericInterface)` method has been refactored. The separate methods `EnsureTypeIsInterface(Type type)` and `EnsureTypeIsGeneric(Type type)` have been removed, and their checks are now performed directly within the `DoesImplementGenericInterface` method. If the `genericInterface` is not a generic interface, an exception is thrown.

Lastly, the `EnsureTypeIsInterface(Type type)` and `EnsureTypeIsGeneric(Type type)` methods have been removed from the `SignificantNumber` struct.