Refactor `SignificantNumber` struct and add new methods
The `CompareTo(object? obj)` method in the `SignificantNumber` struct has been refactored to throw a `NotSupportedException` for all inputs, simplifying the previous implementation. A new generic method `CompareTo<TInput>(TInput other)` has been added to the `SignificantNumber` struct, which is constrained to types that implement the `INumber<TInput>` interface. This method converts the input to a `SignificantNumber` and then compares it to the current instance. Additionally, a new static method `DoesImplementINumber(Type type)` has been added to check if a given `Type` implements the `INumber<>` interface.