Removed MinDecimalPlaces from SignificantNumber struct
The `MinDecimalPlaces` constant was removed from the `SignificantNumber` struct in the `SignificantNumber.cs` file. This change suggests that there is no longer a minimum limit to the number of decimal places that a `SignificantNumber` can have, allowing for more flexibility in the precision of the numbers represented by this struct.