Refactor SignificantNumber initialization and parsing methods
Significant changes have been made to the `SignificantNumber` struct and its methods. The `MaxDecimalPlaces` and `FormatSpecifier` constants were removed. The initialization method for `SignificantNumber` instances has been refactored to use a new method `GetStringFormatForFloatType<TFloat>()` to determine the string format based on the float type. Parsing of the significand and exponent has been updated to check for 'E' in the string representation of the input. Trimming of trailing zeros from the significand now operates on a span instead of a string. The splitting of the significand into integer and fractional components has been updated to handle cases where the significand does not contain a decimal point. A new method `GetStringFormatForFloatType<TFloat>()` has been added to return a string format based on the float type.