Added `To<TOutput>` method to `SignificantNumber`
This commit introduces the `To<TOutput>` method in the `SignificantNumber` class, both in the SignificantNumber.cs file and in the README.md documentation. This method allows conversion of a `SignificantNumber` to a specified numeric type. The README.md file has been updated with a new section on Conversion, providing an example of how to use this new method.

In addition, the constant `ten` in the `SignificantNumber` class has been replaced with `Base10` for improved readability.

New test cases have been added to the AITests.cs file to test the `To<TOutput>` method with various numeric types and to check for overflow exceptions.

The package version has been updated from 1.0.0 to 1.0.1-alpha.1.